translator/requirements.txt
grpcio-tools == 1.30
transformers == 4.15.0
sentencepiece == 0.1.96
